NewGameButton component and styling
